requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.0
psycopg2-binary>=2.9
redis>=5.0
django-cors-headers>=4.3
python-decouple>=3.8
requests>=2.31
gunicorn>=21.2
bleach>=6.0
pytest>=7.4
pytest-django>=4.5
pytest-cov>=4.1
pytest-xdist>=3.3